fix: drop typia LlmJson re-export; document typia-only coerce+validate path
The `export { LlmJson } from "@typia/utils"` dragged `@typia/utils` source into
the `test-typia-exact-optional` workspace's type-check, surfacing pre-existing
`exactOptionalPropertyTypes` violations across `@typia/utils`/`@typia/interface`
(a ripple from the schema-description work) that are unrelated to this change and
out of scope to fix here.

Since every `ILlmFunction` already carries `coerce` and `validate` bound to its
schema, an inliner reaches the correct coerce→validate pipeline with only
`typia` via `func.validate(func.coerce(args))` — no re-export needed.
`LlmJson.validateArguments` stays in `@typia/utils` as the registrar's
single-source helper and a one-call option for `@typia/utils` consumers.

- typia: revert the module re-export (public surface back to master-identical).
- docs (mcp README, mcp.mdx, json.mdx): show the typia-only pattern; frame
  `LlmJson.validateArguments` / `stringify` as `@typia/utils` helpers.
- test: replace the re-export test with a typia-only coerce+validate test.
